make 14.0-RELEASE compatible
Instead of patching the rc.local, I simply just ship a simple one.
It's automated anyway, no need to care if there's dualconsole, etc.
Just fire it away.
Also, make the directory called work.RELVER instead of just work.
Using this scheme it's easier to generate 14.0, 13.2, etc. iso's
staying in the same directory.
If downloading the initial script file fails, the script tries
$maxattempt times.
So far, it's hardcoded at 5.
The problem was with the path anyway. The new fcgiwrap didn't had
the /usr/local/bin directory in the path anymore, so it couldn't
find python.